Correct return type of `getTotalLength` in RNSVGRenderableManager

+++ android/src/main/java/com/horcrux/svg/RNSVGRenderableManager.java
@@ -88,7 +88,7 @@ class RNSVGRenderableManager extends NativeSvgRenderableModuleSpec {
   @SuppressWarnings("unused")
   @ReactMethod(isBlockingSynchronousMethod = true)
   @Override
-  public double getTotalLength(Double tag) {
+  public float getTotalLength(Double tag) {
     RenderableView svg = RenderableViewManager.getRenderableViewByTag(tag.intValue());
     if (svg == null) {
       return 0;
